backend/requirements.txt
asyncpg==0.29.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis[asyncio]==5.0.8
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2

# Import pipeline
ijson==3.3.0
psycopg2-binary==2.9.10
tqdm==4.66.5

# EDDN listener
pyzmq==26.2.0
